fix/authenticator: change AlreadyAuthorised behaviour
Previously authorisation requests for an app that has already
been authorised were declined with an `AlreadyAuthorised` error.
This is an incorrect behaviour at least in two cases:

1. App can lose its credentials (e.g. due to a system or a disk
failure). There was no way to request them again from the same
authenticator because the app is already authorised.

2. A same app can be authorised on different machines. Previous
implementation makes this impossible because of the
`AlreadyAuthorised` response.

This commit changes this behaviour and now authenticator returns
a predictable response regardless of a current app authentication
state.