fix/mock-vault: fix read synchronisation
Previously mock-vault didn't synchronise with the file
system on reads. This approach works if it's confined
to a single process, but if e.g. safe_authenticator
and safe_app are two different processes, then weird
bugs can occur, when safe_app could have a state in
memory that is inconsistent with stored MockVault data.