refactor/all: change network re-connect fns
Now when the network connection is failed, we do not
reconnect to the network automatically - it should be done
by a user. For this purpose several new FFI functions have
been added - app_reconnect and auth_reconnect.

Because of this change we have to store core_tx channel
with the client. That means that Client now has to accept
a type parameter for the message context - so all Client
signatures have been changed to Client<T>.