feat/safe_authenticator: improve recovery for app revocation
This commit fixes the complexity of a case when a user revokes
several apps at once - it will make containers re-encryption a combinatorial
and complex process. This fix streamlines it by using a revocation queue that
is stored in the user's config file, so now only a single app can be revoked
at a time.