fix/nfs: return original file on file_close in read mode
Improve semantics and unify the file_close function behaviour: it returns a
modified File struct if a file was opened in one of the write modes, but
returned ptr::null() if it was opened in the read mode. This inconsistency is
fixed now - we just return an original file struct if a file was opened in
the read mode.

See MAID-2191 for more details.