fix/mock-vault: fix mock vault deserialisation
maidsafe_utilities introduced a new change that turned
out to be a regression: in some cases when a deserialised
struct is serialised again, the binary representation
might be different. This is now an error in the serialisation
functions in maidsafe_utilities because it can prevent tampering
of serialised messages. At the same time, this error prevented
mock-vault from loading successfully in some edge cases.
This commit fixes that by temporarily using bincode to deserialise
the mock-vault data, and now we make sure to output error messages
if there are any (previously they were silenced).