fix/jni: fix classes lookup on Android by caching class loader
By default Android JNI derives the class loader from the last Java class
available in the call stack. This fails in case if we call from non-Java
context (as we do when we call Java callbacks from the event loop thread
in SAFE Client Libs). This commit fixes this by caching the class loader
during the library load and using it afterwards to look up Java classes.