chore/ci: conditionally create package
Without checking for the version-based commit message, the package that
gets produced will always have a version number in it. This is the same
condition that gets checked during the `before_script` step, but it's
easier to just check the same condition again, as it's hard to share
information between steps. We don't need the 2nd condition that appears
in `before_script` because the build will have already failed if that's
wrong.