chore/ci: move packaging to script file
The inline script wouldn't work on AppVeyor because AppVeyor treats
a process printing anything to stderr as a failure. Cargo script
definitely writes to stderr and hence was causing the build to fail. The
issue is documented here:
https://github.com/appveyor/ci/issues/1873

You also need to use the special `PowerShell` notation (as opposed to
`ps`) to get it to work correctly.