chore/ci: provide a rebuild option
This provides a switch that controls whether the script runs a rebuild.
The purpose of the rebuild is to put the libraries into a particular
target directory, which the subsequent `find_libs` function uses.

If there is no rebuild, we just pass `None` to the `find_libs` function,
and it'll just look in the target directory, without appending the
additional search path.