refactor/build: tidy variable usage
COMPAT_DIR wouldn't actually be used in the case where COMPAT_TESTS is
overridden in the environment, so only use it when COMPAT_TESTS hasn't
been provided externally.

Also remove use "On Linux" when referring to building with Docker, as it
can also run on OSX, and possibly Windows too.