feat/build: upload Windows artifacts to S3
The Makefile has a target that packages the artifacts into an archive,
and we then use the Jenkins AWS plugin to upload those artifacts.

The Linux build will be refactored to do this at a later stage.

Some other things were fixed:

* add Z option to volume mount (shared code is not accessible without this option).
* create artifacts directory:
The artifacts need to exist before calling the package-build-artifacts
target. I'm also taking that out of the build-with-container script just
to make it a little more obvious at a glance that the artifacts
directory is created by both Windows and Linux.
* parse mode variable correctly
* real target needs artifact creation