feat/build: run tests on Windows
We can simply just call the 'test-mock' script to run these tests on
Windows. As part of doing this, we need to retrieve the build artifacts
that were made in the previous stage. This may seem unnecessary, but
it's to support a distributed build where each stage could potentially
run on a different machine.

I'm using the Makefile as a point of coordination. This is to support
running the builds somewhere else, like on Travis. Even although we now
know we can't use S3 on Travis, it could be at some later stage we'd
want to run the build on another provider. You can also call these
Makefile targets regardless of the operating system you're running on.

Some other small fixes:

* remove parallel method
* remove previous artifacts directory:
The mkdir artifacts commmand will error if the directory already exists.
We always want a fresh copy of these for each build anyway, so it's fine
to just clear the directory out at the start of the process. It's
important to remove only the artifacts directory and not the target
directory, because for the Windows build that would force a recompile of
all the dependencies.
* retrieve artifacts correctly:
On Linux, the artifacts get shared as the artifacts directory in a
mounted volume in the container. On Windows we need to extract the
artifacts to the target/release directory. Since this process could run
on any Windows slave, we clear out the previous build if they exist
(though it is possible these could be the artifacts we just built if
this stage just happens to run on the same machine as the last stage
which built the artifacts).
* upload binary compatibility tests:
We no longer need the docker cp for this, as the build step already
copied the artifacts out of the container then removed it. We can just
look in the target directory to retrieve them.