feat/build: support builds on OSX
This doesn't really require any significant changes, other than a bit of
logic in the Makefile to say, "if we're running on Linux, use a
container, otherwise just use the build scripts".

It's very cumbersome to use a logical or in the ifeq Makefile syntax
(i.e. to say Windows or OSX), so this way of doing it is easier.