chore/ci: use branch name in artifacts
The build number on its own isn't enough for Multibranch Pipeline jobs.
We need to set the branch name based on whether the CHANGE_ID property
is set. This will be set to the PR reference for PR builds and not set
on branch builds, so we'll just use the branch name for those.