chore/ci: conditional for deployment
Introduces a conditional so that deployment will only occur on a certain
branch. I tried to do this originally by defining the deployment branch
in a parameter, just in case we ever wanted to change that, but doing
the comparison just didn't seem to work correctly. I've tested that this
setup works correctly.