Jenkins Build and Release (#113)
* chore(ci): build in container for linux

Provides a container definition for building safe-cli. Unfortunately
when you then subsequently run another build in the container, it is
performing re-compilation. I've only observed this behaviour on this
project, and don't really know what the reason might be.

There was also a little change here in the testing code, which was using
a reference to safe_cli that was assuming it was in a target directory
that was relative to the root of the repository. For a build in a
container this is not the case, so the code has been changed to check if
the CARGO_TARGET_DIR variable is set, and if so use that, otherwise it
will default to a directory at the root of the repo.

* chore(ci): build and test in jenkins

Provides build and test for all platforms. The build artifacts are
uploaded to S3, with the intention being they will be pulled in for the
deployment stage (which can take place on another machine).

* chore(ci): create tag for build

Creates and pushes a tag for the deployment stage. A tag needs to exist
before a GitHub release can be created.

This way of doing it is unfortunately a bit clunky, but it's the best
option at the moment. See this post on Stackoverflow:
https://stackoverflow.com/questions/33570075/tag-a-repo-from-a-jenkins-workflow-script#

* chore(ci): detect existing tags

It's quite likely when the deployment stage runs that there'll already
be a tag for this version number. Git errors if you try to create a tag
that already exists, so in that case, we just do nothing and move on.

* fix(ci): remove fixuid customisation

This was an attempt to see if you only needed to run fixuid on the main
source directory, but this is not the case.

* chore(ci): delete tag and re-create

If there's a tag for the current version, we'll just delete it and
create a new one with the same version number. The version number will
be incremented manually using semantic versioning, so it's perfectly
possible that there will be a merge to master without a version number
update.

* refactor(ci): change clone to fetch tags

By default the checkout method is not fetching any of the tags for the
repo, which is problematic if we want to check for existing tags.

* fix(ci): attempt to fetch tags

Using the more elaborate checkout method caused Jenkins to fail with
some merging related error (it tries to merge on PR builds).

* fix(ci): get version number

This accidentally got removed in a refactor.

* fix(ci): prevent the build failing

If there are no remote tags that's a scenario that's OK and doesn't need
to fail the build.

* refactor(ci): remove github-release in container

This tool is now installed directly on the build slave and will be
invoked from the Makefile/Jenkinsfile to make it a bit easier to re-use.

* feat(ci): deployment process

The deployment process is as follows:

* Clone the code with the tags fetched (not done by default)
* If the commit message is a version change:
  - Get the version number
  - Package the artifacts with version number in file name
  - Create a tag for this version
  - Publish a GitHub release
* Otherwise:
  - Package the artifacts with commit hash in file name
  - Upload the artifacts to the safe-cli S3 bucket

* fix(ci): test different checkout method

* fix(ci): fetch tags with manual command

This checkout method just doesn't seem to work on this build (though I
have seen it working on another project). It was complaining about
"refusing to merge unrelated histories".

* fix(ci): provide deploy bucket property

* fix(ci): use correct details for release

I had copy and pasted this code from a sample project I was using and
forgot to update various details.

* feat(ci): only deploy on master branch

Now that the process is working, this conditional can be added in. I've
tested this before on other builds, so I know this is the correct
mechanism for testing the branch.

* feat(ci): strip travis

Now that Jenkins covers all platforms, we only need Travis to run Linux
builds.

* style(ci): use java style

Since my main language is Python, I have a habit of using the styles
from that language. I've changed the code to conform with the Java
camelCase styles.

Maybe it's not so much of a big deal in this context, but since the
Groovy standard library functions use the Java style, I'd rather have
the code be consistent with that.

* fix(ci): missing bracket