refactor(ci): new IDs for maidsafe user
This changes the user and main group ID of the maidsafe user to match
the user and group ID of the jenkins user on the docker build slave.
This should avoid the need for fixuid to run.

For the build to work correctly, we need these IDs to match the IDs of
the container host, but thousands of files need to be changed and this
is very slow on an AWS EBS volume. This should avoid the need for the
step and hopefully save a significant amount of time for the build.