chore(ci): rename deploy artifacts
The deployment artifacts are renamed to the standard Rust format for a
'target triple'. These are documented here:

https://github.com/japaric/rust-cross#the-target-triple
https://stackoverflow.com/questions/41742046/is-there-a-list-of-all-cfg-features
https://doc.rust-lang.org/reference/conditional-compilation.html

Self update detects the triple for the platform its running on and then
filters the available releases based on that target. So the filename
attached to the release should contain this target in them.