feat(update): provide an update command
We decided it would be a better user experience if the user could
control the update process rather than having the application check for
an update every time it runs. That check could get annoying, both in
terms of the text output and it might cause a delay of a few seconds
while it goes out to the web to query for new versions.

The main noteworthy thing about this is, since the self_update crate
works with the Error type, but our application uses String for the error
return type, so I'm doing a conversion to a string based error.