refactor(ci): use container with build tag
Using the application version number doesn't actually work very well for
a build container, because when you do a version change, you don't have
the new version of the container available yet. Using a version number
for a container would be more appropriate for distributing the
application after it's already been built.