chore/ci: cache for windows build
Every time a new PR is opened on Jenkins, a new workspace is created.
For this reason, the Windows build needs to compile all the dependencies
from scratch. This change pulls in a cache that is auto generated by
another job.

There are also another couple of changes in here:

* Triggering the cache and docker build jobs after a master build
* Removing macOS from running on Travis