fix/vault: verify if the app is still authorised when reading data
When reading data from the network, we should first check if an app still exists
at the list of auth'ed apps at the source elders. When this app is revoked,
there will be no re-encryption of data. The request will be blocked at the
source elders itself. This check was missing in the mock vault implementation.