chore/ci: build with musl
Building with musl changes the process on Jenkins a bit.

The `cargo test` commands don't work with musl because they use extra
dependencies that won't link correctly, so the approach now is to build
with musl as a separate parallel job; the tests can continue to
execute using the gnu target. The Linux testing job will no longer
upload its artifacts because we're only really interested in the
artifacts from the musl build.

Finally the deployment targets are changed to use 'musl' in the tar file
name because self_udpate will expect it to be in this format.