chore(ci): fix deployment from master
We have to check the 2nd commit for the version change rather than the
first one, because the first will be the automated 'Merge pull
request...' commit.

The only thing is, this doesn't really work too well when you have to
test the deployment processes, but we should be able to work around that
without too much hassle.