feat/update: use self update for new versions
This is the simplest possible implementation of self update. When vault
starts:

* Get the 'target triple' for the currently running binary
* Query the repository to see if any GitHub releases are available for
this target
* If there are releases available:
  - Check for newer version
  - If there's a newer version:
    + Update the version
    + Print a message to tell the user to restart
    + Exit the process
  - Otherwise continue as normal
* Otherwise continue as normal

There will only be ouput on the console if there are updates available.
The user will have to set `RUST_LOG=info` to see messages related to
checking for the updates.

This also changes the CACHE_BRANCH setting, which was wrongly set to
experimental rather than master.