refactor/connection-manager: use a unique quic-p2p instance per connection group
Fixes #993. This approach simplifies the use of different identities with a single vault.
Previously we had to disconnect from the vault if we wanted to use a different identity because
we used a single quic-p2p instance per a ConnectionManager, and hence the same connection/port and
the same challenge response. Changing it to a connection per ConnectionGroup makes it much simpler
because we don't have to do the reconnections anymore.