refactor/ci: new deployment process
This is a new deployment process that checks to see if we have a version
change commit, and if so:

* Package all the artifacts for deployment
* Create a tag at this point in time
* Create a release on GitHub
* Upload the mock versions to S3

If there's no version commit, the artifacts are simply uploaded to S3.

Previously the version check was done in a script, but we're doing that
in the Jenkinsfile now, along with all the other repos.