chore/ci: build iOS mock and real libs
This will build the libraries, package them, then upload them to S3. The
retrieval process has also been extended to include the new artifacts.

The build artifact packaging was extended to include specifying the CPU
architecture and the artifact retrieval has been extended for the new
iOS artifacts.

The `retrieve-build-artifacts` target was removed as this is not used
any more.

The only potentially problematic thing here is we now have 6 jobs that
run on macOS and we have fairly limited resources to service them. They
do run fairly quickly, so hopefully it shouldn't be too much of an issue
in terms of the time it would take to wait on executors becoming
available.