chore/ci: containers for android builds
This introduces new containers for building on Android. The setup for
the armv7 and x86_64 architectures are slightly different, so it's
better to have them both in their own containers to isolate them a bit
more.

The tag for the main container has now been replaced with `x86_64`
rather than the generic `build` tag.