chore/ci: build and package universal ios lib
This part of the build occurs in its own stage because we obviously need
all the individual iOS libraries, including mock and non-mock, to be
built first. It occurs before the deployment stage because it's easier
to retrieve all the artifacts at once during deployment and then just
deploy them, rather than also having to do this.

So all the iOS artifacts that are built in the previous stage are
retrieved here, lipo is used to combine the libraries, then the
universal library is put into a tar and uploaded to S3, to be retrieved
in the deployment stage.