chore/ci: packaging for mobile deployment
The Makefile targets for uploading and retrieving artifacts were getting
a bit out of hand and difficult to maintain. I moved the retrieval out
to a little script that can just go over in a loop and retrieve the
artifacts. However, this involves changing the name of the artifacts to
a known value, and that's been done in this commit too. The Rust 'target
triple' is appended to the end of the name of the archive.

Also shortens the method for using lipo a little bit to make it more
readable and managable.