chore/ci: package mobile libs for deployment
The packaging process for deployment now includes the mobile libraries
for both iOS and Android. I refactored `package.rs` to just allow
passing the target triple directly instead of an indirect architecture
name which didn't really map onto anything and I couldn't see it really
being used for much. The names of the artifacts to be deployed and
attached to the GitHub release now have the proper target triple names
as they're referred to in Rust documentation.

To be honest I'm not entirely sure that after this refactor this script
will continue to build the libraries correctly, if someone still uses
the script for that. For the purposes of the release process we're not
doing a build since we've already done that.

This also introduces parameters for the deployment and publish branches.
They're required to be different because when the release processes are
being developed, you sometimes set the deploy branch to a PR branch, but
you don't want to publish in that case.