refactor/ci: sequential publish to crates.io
Previously we had set this up as a parallel process for each crate, but
it turns out they need to be done sequentially.

This commit also enables a nightly build, and on the nightly we do a dry
run publish to check.

The triggers for kicking off the caches have also been changed to only
run on the nightly build. It's probably a touch excessive to regenerate
the caches on every PR merge.

This also fixes an issue on macOS where the `test-mock` script suddenly
stopped producing the `libsafe_app.a/dylib` files. The solution was to
also call the `build-mock` script, although why it just came up as an
issue now isn't clear.