chore/ci: nightly build with deployment
The nightly build will now deploy the artifacts with `nightly` rather
than a commit hash or the version number. These will go to an S3 bucket
and the previously nightly will be removed each time.

Refactored some of the scripts to make things a bit easier to support
this scenario.