chore/ci: conditionally create github release
We only really want to do a new release if there's a change to either
safe_auth or safe_app. It should be said that a change to *either* of
these will produce a release for both of them. It would be a much bigger
change to only create a GitHub release for these components
individually.