chore(ci): builds for safe-ffi
Currently there are no tests for safe-ffi, so this just does release
builds and adds targets for building containers for ffi.

This also includes ARMv7 and x86_64 builds for Android, with container
definitions for both. There were quite a lot of problems getting ARMv7
to build, in particular it couldn't link libssl.so and libcrypto.so, so
we've provided custom versions of these from this repository:
https://github.com/ph4r05/android-openssl/tree/master/jni/openssl

The artifact packaging is slightly refactored to accomodate the new
mobile platforms.