chore(ci): iOS builds for safe-ffi
Builds safe-ffi for iOS aarch64 and x86_64, and also adds a stage to
combine these libraries into one universal library.

Artifact retrieval is also refactored to use a script as it was
becoming too unwieldy in the Makefile. The script attempts to grab
artifacts for both safe-cli and safe-ffi for all the platform targets.
If there is no artifact for that target for that component, we just move
on to the next target.