chore(ci): mock builds for safe-ffi
Enables mock builds for all the safe-ffi components except for
`x86_64-linux-android`. At the moment there's a linker failure when you
try to build this configuration. There is no problem building the
non-mock configuration, so it's not immediately obvious what the fix is.

This also has a fix for the build-component.sh script to differentiate
between building a binary or a library, based on the name of the
component.

Also fixes a little problem in the retrieve artifacts script: without
this it requires authentication.