chore(ci): publishing for safe-api
Adds a simple target for publishing safe-api, which is the only crate in
this repository that requires a publish.

There's also a refactor in here for the deployment packaging process.
I've refactored this out into a generic script because again there would
just be far too much duplication if I added safe-ffi in the same way.