chore(ci): only clean for prod build
I always like to do a release build (as in something we're going to
release to the public) as a clean build, just to ensure that no issues
with caching creep in, but it's very slow to do both the prod and the
dev builds as clean. Even though we will release the dev builds to the
public, it seems an acceptable risk to not run this as a clean build.