chore(ci): github actions for release process
This transfers the build and release process from Jenkins to GitHub
Actions. The entire process is replicated here, split into 2 workflows, PR and
master.

The PR configuration will run the tests and build the mobile
configurations. The master configuration will build everything, then
deploy and publish. There are conditionals that check for the existence
of a version change message, as before. Since you have the same
configuration for building mobile in both workflows, there's a lot of
duplication here, but for now we're just accepting that as a limitation
of the Actions system.

Due to an issue with rust_sodium, we are also missing macOS for the time
being, but sodium should be removed fairly soon.