chore/ci: use actions for build and release
Uses GitHub Actions for the build and release process, though at this
point, it doesn't contain any builds for macOS, due to an issue with
rust_sodium building on Catalina. The iOS stuff is in here, but it's
currently commented out. When the rust_sodium problem is resolved,
macOS-latest can be added the the appropriate strategy matrix and the
job for building on iOS can be uncommented. After that you'll need to
add a job for building the universal iOS library, but you can copy the
`safe-api` Actions implementation for that.

The strategy matrix is utilised to reduce as much duplication as
possible.

For building on Android, the Makefile is used to help reduce duplication
in the job definitions. Parameters can be passed for the target and the
type of the build.

This also refactors some things around the containers being used:

* All the different targets for building and pushing the container were
replaced with 2 generic targets.
* To make them more easy and generic to use with Actions, the container
tags have been changed to container the build target and type (dev or
prod)

All references to "mock" and "real" should now be removed from the build
and release process, though they are probably still present in other
things for legacy reasons.