docs(UserGuide): add quotes around safe urls with version numbers
Running a safe command which includes a versioned URL on my machine
results in a `No matches for wildcard` error so I'm having to put the
URL in quotes. Maybe this is only macOS, or my specific shell, but
probably does no harm putting the examples in the readme into quotes
as that should work universally.