fix(examples): Migrate routing examples to crate root
It's not possible to use `cargo` to run examples in nested directories,
so the routing examples have been promoted to the crate root. This
required some changes to the examples to make them compatible with the
renamed crate, as well as the addition of `tokio-util` as a
dev-dependency to the root crate.