chore: Remove hand-written `Debug` impls
In almost every case, our hand-written `Debug` impls existed exclusively
to skip/reformat sensitive/overly verbose fields. Recently we added the
`custom_debug` crate to our dependencies, which allows us to make these
tweaks as annotations in the type definition, which is a massive
reduction in boilerplate.

Additionally, this commit takes the somewhat stronger stance that we
should prefer not to customise `Debug` output, except for security
reasons (e.g. don't show private keys) or *extremely* verbose fields
(such as payload or chunk bytes). This is on the basis that, if you are
looking at a type's `Debug` output, you probably want more information,
rather than less.

This will probably mean significantly more verbose `Debug` output, which
is likely to have a detrimental impact on logging clarity. Follow-up
commits will seek to tackle this with the following interventions:

- Be more specific about what to log. E.g., rather than logging an
  entire `Map` struct, just log its address if that's what we care
  about. The idea here is to handle verbosity at the call-site, rather
  than globally reducing verbosity for the type.
- Otherwise, apply verbosity reductions as 'far out' as possible. This
  is likely to affect messaging enums, for example, which sometimes
  contain fairly nested structure which could wind up being particularly
  verbose. Again, the principle would be to make the messaging enum
  responsible for setting its desired verbosity, rather than 'infecting'
  all of its members with those concerns.

In both cases, we should be able to achieve a satisfactory level of
verbosity using `custom_debug`, perhaps with the addition of a couple of
formatting helpers, so we would ideally still have very few hand-written
`Debug` impls.