chore(types): Short `Debug` output for `{PublicKey,Signature}::Ed25519`
`ed25519_dalek` uses the default derived `Debug` output, which is very
verbose for these types as they contain large byte arrays. It's also
inconsistent with the other variants, from `blsttc`, which use an
abbreviated slice of the byte representation. `custom_debug` has been
used to apply the abbreviated format to the `Ed25519` variants.