ci: Build and Deploy for ARM
Provides builds for ARM and ARMv7 and also updates the deploy process
to include those builds. The builds for ARM are performed using the
`cross` tool, which sets up the context for building inside a Docker
container. This wraps the process quite nicely and prevents you having
to go to any trouble of getting all the required environment variables
correct.

For the PR workflow, a job has been added to make sure that the ARM
architectures can build OK.

It didn't seem valuable to add tests specifically for the ARM
architecture, since the tests that currently run for this repo seem
like they're more functional.

Finally, verbose logging for Cargo was removed, on the basis that we
agreed it wasn't very useful.