feat(messaging): Add a `ClientAuthority` 'proof' struct
The `ClientAuthority` struct can only be constructed by verifying a
`ClientSigned` for a particular payload. Thus, instances of
`ClientAuthority` can be trusted as proof that the payload indeed had a
valid signature, and was signed by the owner of `public_key` (e.g. a
holder of the corresponding secret key).

Note that it's important that `ClientAuthority` **does not** implement
`Deserialize`, or any other trait that could allow unverified
construction. It therefore cannot be a drop-in replacement for
`ClientSigned`, which is kind of the point as we can now introduce
`ClientAuthority` in areas where we expect the signature to be valid,
then work our way back towards verification. Ideally we should be able
to verify immediately and use `ClientAuthority` in `MessageType`.