fix(node): Ensure storage write op authority is verified
Previously, we would only verify the signatures of the original client
message, not the signatures of replicated ops. With this change, a
`ClientAuthority` is now needed to apply an op, meaning the signature
needs to be verified in order to call those functions.

Currently, verifications happens quite late, inside each individual
store. This is because the `DataExchange` content is not iterated until
this point. We really want to do this verification earlier, however.