fix(deploy): Only Package Binary File
There was a little bug in the deployment process that was introduced by
the use of a wildcard. This was causing extra files to be picked up and
packaged in the archives. The reason for using the wildcard was because
of the different name of the binary for the Windows build.

An additional line has been added to the script to specify the name of
the binary, based on whether we're packaging for Windows or not.

Removing the wildcard meant the tar process could go back to the
original way it was being done without the subshell, which is a bit
more readable.

There's also a little utility target added to the Makefile for getting
the artifacts setup in the same way as they are during the build
process. This just makes it easier to debug the packaging process
locally.

Also added the zip files for the ARM builds to the release. For some
reason that was missed before.