refactor(messaging): Remove message info from data write ops
BREAKING CHANGE: The `*Write` structs contain the information needed to replay a cmd, the
`write` op itself, and the `client_sig` authorising it. They also
contained message info, the `msg_id` and `origin` of the message that
contained the op.

The message info is used in the initial call to `store.write` to
construct `CmdError` replies for the origin. It was largely unused in
the replication logic, except to add the message ID to some logged
errors (specifically when trying to delete a public register or
sequence). It therefore makes sense to remove these fields from the
replicated ops.