fix(node): Stores always verify client authority
We use `ClientAuthority` in `NodeDuty` to reflect the fact that, by the
time we have constructed a `NodeDuty`, we expect the client authority to
be valid. Note that authentication is still a separate step, but the
public key in the authority can be trusted to have signed the op.

Ultimately, this makes it impossible to call any of the store functions
without a `ClientAuthority`, which can only be constructed by verifying
a `ClientSigned` for the op.