ci: Deployment Test
Adds a utility script for checking that a deployment has all the
correct details as we would expect.

It uses the very handy Github CLI tool to get the details and URLs
for all the assets on the release, which means you can download and
verify them quite easily.

This script will help when changes are made to the deployment, as I had
inadvertently broke things when I last made a change.