refactor(types)!: More compact ed25519 signature serialisation
`ed25519_dalek::Signature` serializes as a 64 element tuple, which
`rmp_serde` serializes as a regular array in which each byte is
serialized individually. Concretely, this meant any byte larger than
0x7f/127 would be serialized as *two* bytes (one tagging it as a uint 8,
and the second for the actual byte).

Message Pack does define a byte array type, which serializes as a tag,
length, and bytes. `rmp_serde` supports this through serde's `*_bytes`
methods, which we can use `serde_bytes` to hook into.

BREAKING CHANGE: Although the API doesn't change, signatures are
serialized in messages, so this break the message format.