refactor(messaging)!: Add an `Authority` 'proof' struct
This generalises the approach taken with `ClientAuthority` of wrapping a
`*Signed` struct in such a way that they can only be constructed
(outside the `authority` module) with a valid signature for a particular
payload.

It has the same advantage as `ClientSigned`: an `Authority<NodeSigned>`
is guaranteed to contain a valid `NodeSigned`. By extension, a
`NodeSigned` could be assumed to be unverified.

This was much easier to 'plumb in' than `ClientAuthority`, since it's
used in fewer places. As with `ClientAuthority`, signatures would
currently be double-verified (once by `WireMsgUtils::check_signature`
and once again in `WireMsg::into_message`). This can change once all
`*Signed` types have a corresponding `*Authority`.

BREAKING CHANGE: Some public types that contained `NodeSigned` now
contain `Authority<NodeSigned>`. Usage may require unwrapping the value
with `into_inner`.