refactor(node): Use Github Releases for Install
An issue was discovered with the S3 provider for the self_update crate
where it wouldn't reliably return the latest version of the sn_node
release. This is because it uses the S3 Rest API and only returns 100
items from the bucket being queried. The order of items returned by the
API are based on the 'key' of the item, which is a string value, so
sometimes it would return the latest version and sometimes it wouldn't.

The code has now changed to use Github as the provider, which is not
subject to this problem. The self_update crate also had to be changed
to support this, as it turned out it wasn't actually parsing the
release assets correctly. For now this commit references this specific
version of the self_update crate, which hasn't been merged yet.

A test has been added to the network tests for testing the `safe node
install` command. The test builds the CLI from the current commit
and then runs the `safe node install` command from that build, and
asserts that it has installed the latest version.