refactor(messaging)!: Use `Authority<SectionSigned>`
We now use `Authority<SectionSigned>` wherever we expect a valid
`SectionSigned`. As with `NodeSigned`, this was quite straightforward to
plumb in.

BREAKING CHANGE: Some public types that contained `SectionSigned` now
contain `Authority<SectionSigned>`. Usage may require unwrapping the
value with `into_inner`, or using the new getters to read fields.