refactor(node): Support New Arch for Linux
It turned out that the code doing a check for the `target_os` wasn't
enough, as that is still true for the Aarch64 build, but the
target_arch is different. The current code was then pulling in the
x64 Linux build when running on Arch64.

I've changed the code to replace "gnu" with "musl" for any type of
Linux build, since this is currently true for all of them. We'd need
to update this code if we produce a Linux build that's not using musl.