refactor(messaging)!: Make header size exclusive of version and size
This is another step towards simplifying message serialization with
bincode. This brings us closer to the `(size, bytes)` portion of the
message being directly serializable as `&[u8]`, since `size` needs to be
the size of bytes, rather than the whole container.

The `WireMsgHeader::write` method still returns the overall serialized
size, so the interface with `WireMsg` doesn't need to change.

BREAKING CHANGE: This breaks the wire message format. Messages from
earlier versions will not be readable in this version (and vice-versa).