refactor(messaging): Derive most of `WireMsgHeader` serialization
With the previous changes, it's now possible to (de)serialize the
message header as a struct containing the version and the payload bytes
(which would be encoded with the length as a prefix). This has been
called `MessageSkeleton`, for want of a better name, but since it's
private to the module it's not too important.

The unchanged test shows that this is compatible with the previously
hand-written serialization code.