refactor(messaging): Remove `u16` limit from header size
This probably existed in order to have a fixed-size encoding, and `u16`
is a pretty sensible limit for a header size. However, since we can now
support arbitrary `usize` through bincode, and since message size limits
would be better handled earlier in the stack, we can simplify things by
removing this limit.