chore: Update qp2p to get more connection retries
Hopefully this will make the tests a bit less flaky, though the jury's
still out.

This was also a breaking change for qp2p, but released in a patch
version. Technically though the semver spec says the only rule for
pre-1.0 is "there are no rules". This did require removing a use of the
removed field.