feat(node): Support Install of Specific Version
The `safe node install` command is extended to support installing a
specific version of the node. This is useful for testing scenarios
where the client needs to be compatible with a specfic version.

A test was also added to the network tests to check that the command
works as expected with the `--version` argument.