refactor!: Remove `testing` as a default feature
This probably should not be default, as downstream consumers who want to
do testing can easily turn on the feature in their own dependencies.

This does require explicitly turning on the feature for the benches and
examples that we need it for, and explicitly turning it on when
overriding features in workflows.

BREAKING CHANGE: This removes items from the (default) public API. They
can be re-added by enabling the `testing` feature for the `safe_network`
dependency.