test: CLI and Network Tests
This is a set of tests that test various aspects of the CLI, nodes, and
a testnet. The suite was originally written to test that the ARM builds
were functioning correctly, but with a little bit of extension, they
will be able to run against any build.

The tests are as follows:

The Safe Network CLI should build successfully
The Safe Network CLI install script should run successfully
The Safe Network CLI should be installed at the correct location
The Safe Network CLI install script should install an ARM build
The Safe Network CLI install script should install the correct version
The Safe Network CLI update should run successfully
The Safe Network CLI update should update to the latest version
The Safe Network CLI should successfully install a node
The Safe Network CLI should install the node with the correct version
The Safe Network CLI should successfully update a node
The Safe Network CLI node update should update the node to the latest version
The Safe Network CLI should successfully add the local node to a testnet
The Safe Network CLI should successfully create keys on the testnet
The Safe Network CLI should successfully upload a file to the testnet
The Safe Network CLI should successfully download a file from the testnet

It's quite a large script and arguably getting to the stage where it
should be implemented in something like Python rather than Bash, but it
should hopefully be pretty maintable, as everything is broken up into
functions (which is probably how it would be written in Python anyway).
Python is a little bit more cumbersome for getting output from
processes, which is why I ultimately decided to just leave it in Bash,
as this script does a lot of that.