refactor(client)!: Construct `Client` with `Config`
This changes the signature of `Client::new` to take a `Config` instance
rather than separate optional `config_file_path` and `bootstrap_config`.
This is a better separation of concerns (e.g. set up your configuration,
then constuct a client with that configuration) and will facilitate
further improvements to this API.

BREAKING CHANGE: The signature for `Client::new` has changed to require
a `Config`. `Config` itself can be constructed from the arguments that
used to be supplied to `Client::new`.