refactor(client)!: Re-export `Config` from `client`
With `Config` becoming pat of the interface to construct a `Client`, it
would be nicer for it to be exported at the same point. Furthermore,
`Config` is the only public item in `client::config_handler`, so that
module has been made private.

BREAKING CHANGE: Client `Config` now needs to be imported from
`client::Config`, rather than `client::config_handler::Config`.