fix(client): Don't set `forward_port` for clients
Clients do not need a reachable public IP, so turning this on just gives
us (at least) an additional round trip for now benefit.

This will also allow us to drop the `client` feature as it is no longer
needed.

Note that this could be regarded as a breaking change since clients
would no longer be reachable by a public IP, but we expect this not to
be a problem in practice.