test: Normalise features in E2Es
All compiling in E2E tests now uses both the `always-joinable` and
`testing` features. In some cases this is adding the `testing` feature,
which would only introduce extraneous items. Specifically:

- All `cargo` invocations in e2e.yml use the same features.
- `testnet`, which builds `sn_node`, uses the same features used for it.
- `network_split`, which is only run from e2e.yml, uses the same
  features as e2e.yml.

This should reduce the amount of compiling we have to do in E2Es. The
cache key has been tweaked as well to ensure a fresh cache is generated
with these changes.