feat(install): Support Specific Version
The install script has been extended to support passing a specific
version number.

It would have been nice to wrap the argument parsing in a little function, but
unfortunately it seems to be the case that you can't consume "$@" inside a
function.

The reason for doing this manually and not using getopts or getopt, is because
macOS doesn't have a getopts install and the version of getopt it has by
default is not the GNU one, so the behaviour will be different on macOS and
Linux. For that reason, we just parse them 'manually'.

This Stackoverflow post has the details:
https://stackoverflow.com/questions/402377/using-getopts-to-process-long-and-short-command-line-options

Tests have been added for this to run on both Windows and macOS, since
we'll want to make sure the functionality works on both.

This also fixes a mistake with using the Rust hash in the caching.