refactor(messaging): Use serde for wire message metadata
Our wire message header begins with some metadata: the total header
length (inclusive of the metadata itself) and the protocol version.
These are (de)serialized with `(from_)to_be_bytes`.

This is exactly the format that `bincode` would serialize a struct of
two `u16`s, when configured to use big-endian and fixed-length encoding.
So, we can replace our hand-crafted, artisinal serialization code with
more straightforward serde calls, whilst maintaining the exact same
format.

Overall this is more code, due to the `HeaderMeta` struct and
`BINCODE_OPTS`, but the `from_bytes` and `write` methods are much
simpler.