refactor: Use color_eyre for Error Handling
This replaces the use of the anyhow crate with color_eyre/eyre.

There are a lot of files changed, but it's essentially just a few
changes:

* Swap the anyhow::bail macro for the eyre::bail macro
* Swap the anyhow::anyhow macro for the eyre::eyre macro
* Swap the context function for the color_eyre wrap_err function
* Swap the with_context function for the color_eyre wrap_err_with
  function

So functionally it's mostly the same, but we get coloured output from
the errors. However, we can start to add better tracing and supplement
the errors using color_eyre's suggestion function.