tests: Update Tests to Catch Up with App Changes
There were several changes required to bring the tests up-to-date with
changes that were made in the application. Of all the tests that are
now enabled to run, I've seen them complete at least once. However, if
you try to run the whole suite, a significant percentage of the tests
will still fail sporadically.

There are a few tests that have been identified that fail consistently:
those have been marked with ignore attributes and need to be fixed.

The main, notable changes made for catch up are as follows:

* Remove the use of the 'safe://' prefix in NRS strings.
* Small changes in the format of error messages from the safe API.
* For files based tests, contents of test data was updated and hard
  coded values for things like sizes were changed to read the test data
  dynamically.
* For files and a few other test categories, the use of the duct cmd!
  macro has been replaced in favour of the safe_cmd helper function,
  which will output stdout and stderr for failed commands.
* Wallet tests and benchmarks completely removed since commands related
  to wallets have been removed.
* Tests related to the `keypair`, `keys balance` and `keys transfer`
  have also been removed, since these commands no longer exist.

The tests have also been configured to use Color Eyre, which is useful
for more detailed error output.